+++ sites/msb/components/static/search/CoreSearchPagination.tsx
@@ -1,5 +1,6 @@
 import { Pagination } from '@matsugov/ui/Pagination';
 import Link from 'next/link';
+import { useEffect } from 'react';
 import { PaginationProps, usePagination } from 'react-instantsearch';
 
 export function CoreSearchPagination(props: PaginationProps) {
@@ -10,6 +11,10 @@ export function CoreSearchPagination(props: PaginationProps) {
   function handleLinkCLick(page: string | number) {
     refine((typeof page === 'number' ? page : parseInt(page)) - 1);
   }
+
+  useEffect(() => {
+    window.scrollTo({ top: 0, behavior: 'instant' });
+  }, [currentRefinement]);
 
   return (
     <Pagination